Initial commit of timer
A category-based timer tracker that uses special machine tags
associated with tiddlers to keep track of time.

  timer:<category>-{start,stop}

The title of the tiddler is a timestamp. text is used for
(optional) annotations.

Still rudimentary, but the idea is that on a phone or similar
it will be possible to see, at a glance, if you are currently
working at a particular category, and start or stop as needed.

There will also be a reset button which retags all the current
category's tiddlers to timer-<timestamp>: for archiving. But
that's not done yet.